python-backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.9.0
httpx>=0.27.0
python-dotenv>=1.0.1
sqlmodel>=0.0.22
pymysql>=1.1.1
asyncmy>=0.2.11
python-multipart>=0.0.12
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
